removing api and making use of auth
